django_dev/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
asgiref==3.11.0
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
psycopg[binary]
gunicorn
whitenoise